requirements.txt
fastapi
uvicorn[standard]
openai==1.*
elevenlabs==1.*
twilio>=9
python-dotenv
websockets
pydub        # converte áudio (pcm ↔ wav)